Algorithm Rework
Originally the algorithm, upon creating events for each day, created an object of all available time windows for the day, based on its constraints and existing events.
The problem with this approach is that when filling out events for a day, the algorithm reached a point where it didn't know what events already existed (because it viewed the day just as available time windows).

This limits its growth. For example, what if you want to ignore specific events, or not place breaks around constraints, or customize the project later on and take into consideration existing events?

In the new version of the algorithm it doesn't perform that breakdown into available time windows. It makes much of our old code useless now, but the algorithm should be more adaptable now too.