Enhanced project management page
Improved the project management page. Each project displays some calculations (events done, time left, etc.), and there's the basis for editing the project details in place: 3 buttons (edit, cancel, save), and started moving the data to input fields.

Later on we need to save the changes in the DB, and most importantly allow creating new events to the project from this page.

Right now all the changes are on the client side only (and without even affecting the schedule yet, for example with name change).