Working algorithm
Created the basis for rescheduling an event.
It required minor changes to the algorithm, with an auxiliary, preparation function. What it does is minor manipulations on the project's parameters. specifically it's time estimate, which is set to be the reschedule event's duration, and the start time for the project, which is set to be either the current date or the rescheduled event's date if it's in the future.
Note that these changes easily allow for us to expand this feature in the future to reschedule multiple events, not just one.

What is missing is presenting the suggestions at the front end.