Daily Time Window
Project parameters now allow defining the daily time window for the project. Meaning - the time window in which the algorithm will search to place its sessions in. By default the time window is set to be all day (from 00:00 to 23:59), but the user can change it. For example if he wants to work on the project he's defining only at evening he can state in the time window 18:00-20:00.