Synced Google calendars of shared project
When all users export their shared project to Google calendar, they share the same schedule and the sync occurs naturally through Google.
The first person to export creates the calendar and creates an ACL for all the participating users, so that when they want to export the project (that has already been exported), they just add the calendar that has been created to their own calendar list.

Left to do: Sync in a mixed scenario, where some have exported, and some have left it local.